requirements.txt
# Core Dependencies for ETL
redis
pyspark
pandas
pyarrow
numpy

# Database
psycopg2-binary==2.9.9

# Utilities
python-dotenv
click

# Apache Airflow
apache-airflow[celery]==2.8.1
apache-airflow-providers-apache-spark